+++ src/infra/src/70_domains/onemail_app/99_data.tf
@@ -12,6 +12,13 @@ data "aws_vpc_endpoint" "dynamodb" {
 
 data "aws_vpc_endpoint" "api_gtw" {
   service_name = "com.amazonaws.eu-south-1.execute-api"
+  vpc_id       = data.aws_vpc.core.id
+}
+
+data "aws_acm_certificate" "api_custom_domain" {
+  domain      = local.zone_name
+  statuses    = ["ISSUED"]
+  most_recent = true
 }
 
 data "aws_lb" "nlb" {
@@ -54,8 +61,20 @@ data "aws_route53_zone" "onemail" {
   private_zone = false
 }
 
+data "aws_ses_domain_identity" "onemail" {
+  count  = var.enable_ses ? 1 : 0
+  domain = local.zone_name
+}
+
 
 data "aws_lb_listener" "ecs_core" {
   load_balancer_arn = data.aws_lb.nlb.arn
   port              = 3000
 }
+
+data "aws_sesv2_configuration_set" "oml_config_set" {
+  count                  = var.enable_ses ? 1 : 0
+  configuration_set_name = "${local.project_nodomain}-${var.env}-configuration-set"
+}
+
+data "aws_caller_identity" "current" {}