+++ src/onemail/om-common/src/utils/constants.ts
@@ -1,0 +1,4 @@
+export enum SES_SIMULATOR {
+  BOUNCE = 'bounce@simulator.amazonses.com',
+  SUCCESS = 'success@simulator.amazonses.com',
+}