+++ src/onemail/om-ecs-dispatcher/src/dtos/email/emailHighPriority.dto.ts
@@ -22,7 +22,7 @@ export const TemplateContentSchema = z.object({
 export const EmailContentSchema = z.object({
   subject: stringCheckedSchema().describe('Subject of the email'),
   html: htmlContentSchema.describe('HTML content of the email'),
-  text: stringCheckedSchema({ max: 200000 })
+  text: stringCheckedSchema({ max: 50000 })
     .optional()
     .describe(
       'Text content of the email used as fallback if HTML content is not supported by the email client',
@@ -31,7 +31,9 @@ export const EmailContentSchema = z.object({
 
 const BaseEmailSchema = z.object({
   from: EmailAddressSchema.describe('Sender of the email'),
-  to: EmailAddressSchema.describe('Recipient of the email'),
+  to: EmailAddressSchema.describe(
+    'Recipient of the email, ignored in case of dryRun',
+  ),
   extendedHeaders: ExtendedHeadersSchema.optional(),
   tag: TagSchema.optional(),
 });