+++ src/infra/src/70_domains/onemail_app/CHANGELOG.md
@@ -1,5 +1,59 @@
 # infra-domains-onemail-app
 
+## 1.7.1
+
+### Patch Changes
+
+- 7a2ec16: user http_proxy instead http for integration requests
+
+## 1.7.0
+
+### Minor Changes
+
+- 4e6a558: Lambda Sender egress rule
+
+## 1.6.0
+
+### Minor Changes
+
+- 0c893fd: SES VPC endpoint and fix Lambda permissions
+
+## 1.5.0
+
+### Minor Changes
+
+- 00e8135: KMS permissions ECS and Lambda
+
+## 1.4.0
+
+### Minor Changes
+
+- 0240cd5: Adding kms permissions for ecs and lambda
+
+## 1.3.0
+
+### Minor Changes
+
+- e74abf3: AWS SES configuration and DNS records creation
+
+## 1.2.0
+
+### Minor Changes
+
+- e19ea7d: Lambda and Ecs dynamodb permissions
+
+## 1.1.1
+
+### Patch Changes
+
+- fe201cd: update terraform lock with linux platform and dx dev container template
+
+## 1.1.0
+
+### Minor Changes
+
+- 1e8b22d: ApiGtw custom domain
+
 ## 1.0.1
 
 ### Patch Changes